refactor: relocate golden and panel harnesses outside support tree (#4117)
Move golden_capture, production/technology panel, screen-spec, leader-selection,
and dark-token assertion modules to app/test/ with shrink-only re-export stubs;
ratchet appTestSupportLocCeiling 962 → 289.
